Added removeItem functionality into Shopping Cart
The "Remove from cart" buttons in the shopping cart now work. In order to
complete the functionality, I had to remove some flexibility. A user is now
only able to add one of each item into the cart. An idea to fix it would be to
add a quantity property to the cart items and display that quantity in the
shopping cart.